+++ sel/src/Sel/HMAC/SHA512_256.hs
@@ -80,6 +80,7 @@ import LibSodium.Bindings.SHA2
   )
 import LibSodium.Bindings.SecureMemory (finalizerSodiumFree, sodiumMalloc)
 import Sel.Internal (allocateWith, foreignPtrEqConstantTime, foreignPtrOrdConstantTime)
+import Sel.Internal.Sodium (binaryToHex)
 
 -- $introduction
 -- The 'authenticate' function computes an authentication tag for a message and a secret key,
@@ -315,14 +316,14 @@ unsafeAuthenticationKeyToBinary (AuthenticationKey authenticationKeyForeignPtr) 
       (Foreign.castForeignPtr @CUChar @Word8 authenticationKeyForeignPtr)
       (fromIntegral @CSize @Int cryptoAuthHMACSHA512256KeyBytes)
 
--- | Convert a 'AuthenticationKey to a hexadecimal-encoded 'StrictByteString'.
+-- | Convert a 'AuthenticationKey to a hexadecimal-encoded 'StrictByteString' in constant time.
 --
 -- ⚠️  Be prudent as to where you store it!
 --
 -- @since 0.0.1.0
 unsafeAuthenticationKeyToHexByteString :: AuthenticationKey -> StrictByteString
-unsafeAuthenticationKeyToHexByteString =
-  Base16.extractBase16 . Base16.encodeBase16' . unsafeAuthenticationKeyToBinary
+unsafeAuthenticationKeyToHexByteString (AuthenticationKey authenticationKeyForeignPtr) =
+  binaryToHex authenticationKeyForeignPtr cryptoAuthHMACSHA512256KeyBytes
 
 -- | A secret authentication key of size 'cryptoAuthHMACSHA512256Bytes'.
 --
@@ -354,14 +355,12 @@ instance Ord AuthenticationTag where
 instance Show AuthenticationTag where
   show = BS.unpackChars . authenticationTagToHexByteString
 
--- | Convert an 'AuthenticationTag' to a hexadecimal-encoded 'StrictByteString'.
+-- | Convert an 'AuthenticationTag' to a hexadecimal-encoded 'StrictByteString' in constant time.
 --
 -- @since 0.0.1.0
 authenticationTagToHexByteString :: AuthenticationTag -> StrictByteString
-authenticationTagToHexByteString authenticationTag =
-  Base16.extractBase16 $
-    Base16.encodeBase16' $
-      authenticationTagToBinary authenticationTag
+authenticationTagToHexByteString (AuthenticationTag authenticationTagForeignPtr) =
+  binaryToHex authenticationTagForeignPtr cryptoAuthHMACSHA512256Bytes
 
 -- | Convert an 'AuthenticationTag' to a binary 'StrictByteString'.
 --